+++ src/roiextractors/extractors/numpyextractors/numpyextractors.py
@@ -24,7 +24,6 @@ class NumpyImagingExtractor(ImagingExtractor):
     """An ImagingExtractor specified by timeseries .npy file, sampling frequency, and channel names."""
 
     extractor_name = "NumpyImagingExtractor"
-    installed = True
     is_writable = True
     installation_mesg = ""  # error message when not installed
 
@@ -247,6 +246,11 @@ class NumpyImagingExtractor(ImagingExtractor):
         overwrite: bool
             If True, overwrite file if it already exists.
         """
+        warn(
+            "The write_imaging function is deprecated and will be removed on or after September 2025. ROIExtractors is no longer supporting write operations.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         save_path = Path(save_path)
         assert save_path.suffix == ".npy", "'save_path' should have a .npy extension"
 
@@ -268,7 +272,6 @@ class NumpySegmentationExtractor(SegmentationExtractor):
     """
 
     extractor_name = "NumpySegmentationExtractor"
-    installed = True  # check at class level if installed or not
     is_writable = True
     mode = "file"
     installation_mesg = ""  # error message when not installed
@@ -470,6 +473,11 @@ class NumpySegmentationExtractor(SegmentationExtractor):
         -----
         This method is not implemented yet.
         """
+        warn(
+            "The write_segmentation function is deprecated and will be removed on or after September 2025. ROIExtractors is no longer supporting write operations.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         raise NotImplementedError
 
     # defining the abstract class informed methods: