+++ src/roiextractors/extractors/schnitzerextractor/cnmfesegmentationextractor.py
@@ -7,6 +7,7 @@ CnmfeSegmentationExtractor
 """
 
 from pathlib import Path
+from warnings import warn
 
 import h5py
 import numpy as np
@@ -147,6 +148,11 @@ class CnmfeSegmentationExtractor(SegmentationExtractor):
         AssertionError
             If save_path is not a *.mat file.
         """
+        warn(
+            "The write_segmentation function is deprecated and will be removed on or after September 2025. ROIExtractors is no longer supporting write operations.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         save_path = Path(save_path)
         assert save_path.suffix == ".mat", "'save_path' must be a *.mat file"
         if save_path.is_file():