+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ### Features
 * New `read_scanimage_metadata` for reading scanimage metadata from a file directly as a python dict [#405](https://github.com/catalystneuro/roiextractors/pull/401)
-* Added `ScanImageImagingExtractor` for simplifying reading ScanImage data [#412](https://github.com/catalystneuro/roiextractors/pull/412) [PR # 417](https://github.com/catalystneuro/roiextractors/pull/417)
+* Added `ScanImageImagingExtractor` for simplifying reading ScanImage data [#412](https://github.com/catalystneuro/roiextractors/pull/412)
+* Added support for multiple samples per slice to `ScanImageIMagingExtractor` [PR # 417](https://github.com/catalystneuro/roiextractors/pull/417)
+* Added volumetric imaging support with `is_volumetric` flag, `get_frame_shape`, `get_num_planes`, and `get_volume_shape` methods [PR #418](https://github.com/catalystneuro/roiextractors/pull/418)
 
 
 ### Fixes